daemon: move protocol files into subfolder
